Pretty large change to WebSeer
Note, if you were attempting to use this plugin, you will need to
re-install.  There are many fixup's in this commit, but the plugin is
still not fully operational.  Email support is still not functioning
correctly.  Also, the concept of the remote.php for remote servers to be
able to add service checks is not tested.  However, the schema and most
of the GUI interactions have been updated.